scraper/requirements.txt
python-dotenv==1.0.0
selenium==4.15.2
supabase==2.0.0
